Added transitive op implementations for MaybeOwnedMut. (#11)
The orginal bounds on the operatores required `&mut A`
references but in manny places this doesn't make sense and
`&A` should be used, even in a MaybeOwnedMut.

For example in `let b = a1 + &mut a2` the variable a2
should never be changed even through you add `&mut a2`,
as this would not be how `+` normally work. I.e. we do
not want to support side-effects in `+`. Additionally
most types implementing Add and similar do so for the
type and the reference to it but not for a mutable
reference to it. Note that *Assign still does use a
`&mut` if available, but only for the left hand side.